+++ core/model/filters.go
@@ -27,6 +27,7 @@ type GroupsFilter struct {
 	IncludeHidden    *bool                          `json:"include_hidden"`     // Include hidden groups
 	Hidden           *bool                          `json:"hidden"`             // Filter by hidden flag. Values: true (show only hidden), false (show only not hidden), missing - don't do any filtering on this field.
 	ExcludeMyGroups  *bool                          `json:"exclude_my_groups"`  // Exclude My groups
+	AuthmanEnabled   *bool                          `json:"authman_enabled"`
 	ResearchOpen     *bool                          `json:"research_open"`
 	ResearchGroup    *bool                          `json:"research_group"`
 	ResearchAnswers  map[string]map[string][]string `json:"research_answers"`