+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Ability to filter authman/manged groups [#441](https://github.com/rokwire/groups-building-block/issues/441)
+
 ## [1.37.1] - 2024-04-24
 ### Fixed
 - Additional fix for scheduled posts[#437](https://github.com/rokwire/groups-building-block/issues/437)