+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
+- Implement create and update post admin APIs [#448](https://github.com/rokwire/groups-building-block/issues/448)
+- API for creating Group for the Admin app [#445](https://github.com/rokwire/groups-building-block/issues/445)
+- API for updating group for the Admin app [#446](https://github.com/rokwire/groups-building-block/issues/446)
 - Ability to filter authman/manged groups [#441](https://github.com/rokwire/groups-building-block/issues/441)
 
 ## [1.37.1] - 2024-04-24